+++ src/games/Checkers/constants/GameSettings.ts
@@ -1,5 +1,12 @@
+// Grid setup
 export const GRID_CELL_WIDTH = 100;
 export const GRID_COLUMNS = 8;
 export const GRID_ROWS = 8;
 export const GRID_WIDTH = GRID_CELL_WIDTH * GRID_COLUMNS;
+
+// Team setup
 export const TEAM_NAMES = ["Red Team", "Blue Team"];
+export const MAX_TEAMS = 2;
+export const EMPTY_CELL = -1;
+export const TEAM_0 = 0;
+export const TEAM_1 = 1;
